--- src/main/java/InvalidCommandException.java
@@ -1,5 +1,0 @@
-public class InvalidCommandException extends InvalidInputException {
-    public InvalidCommandException() {
-        super("I don't know what that means!!!\nEnter \"help\" for a list of what I do know! rawr");
-    }
-}
