--- src/main/java/InvalidInputException.java
@@ -1,5 +1,0 @@
-public class InvalidInputException extends Exception{
-    public InvalidInputException(String message) {
-        super(message);
-    }
-}
