--- src/main/java/InvalidTaskException.java
@@ -1,5 +1,0 @@
-public class InvalidTaskException extends Exception {
-    public InvalidTaskException() {
-        super();
-    }
-}
