--- src/main/java/ToDo.java
@@ -1,16 +1,0 @@
-public class ToDo extends Task {
-
-    public ToDo(String description, boolean isDone) {
-        super(description, isDone);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-
-    @Override
-    public String formatted() {
-        return "T | " + super.formatted();
-    }
-}
