+++ src/main/java/rex/Rex.java
@@ -1,3 +1,9 @@
+import exceptions.InvalidInputException;
+import tasks.*;
+import util.Parser;
+import util.Storage;
+import util.Ui;
+
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
