+++ src/main/java/rex/exception/InvalidCommandException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class InvalidCommandException extends InvalidInputException {
+    public InvalidCommandException() {
+        super("I don't know what that means!!!\nEnter \"help\" for a list of what I do know! rawr");
+    }
+}
