+++ src/main/java/rex/exception/InvalidInputException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class InvalidInputException extends Exception{
+    public InvalidInputException(String message) {
+        super(message);
+    }
+}
