+++ src/main/java/rex/task/Deadline.java
@@ -1,0 +1,27 @@
+package tasks;
+
+import tasks.Task;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class Deadline extends Task {
+    private LocalDateTime by;
+
+    public Deadline(String description, boolean isDone, LocalDateTime by) {
+        super(description, isDone);
+        this.by = by;
+    }
+
+    @Override
+    public String toString() {
+        DateTimeFormatter outputFormat = DateTimeFormatter.ofPattern("dd MMM yyyy HHmm");
+        return "[D]" + super.toString() + "(by: " + by.format(outputFormat) + "HRS)";
+    }
+
+    @Override
+    public String formatted() {
+        DateTimeFormatter fileFormat = DateTimeFormatter.ofPattern("dd-MM-yy HHmm");
+        return "D | " + super.formatted() + " | " + by.format(fileFormat);
+    }
+}
