+++ src/main/java/rex/task/ToDo.java
@@ -1,0 +1,20 @@
+package tasks;
+
+import tasks.Task;
+
+public class ToDo extends Task {
+
+    public ToDo(String description, boolean isDone) {
+        super(description, isDone);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+
+    @Override
+    public String formatted() {
+        return "T | " + super.formatted();
+    }
+}
